cleanup map hooks
